change to path-based params instead of query string and rename file to be consistent with bookCovers function
